refactor stepids
