+++ engine/steps/add-asset/test/test.ts
@@ -1,7 +1,8 @@
 import { expect } from 'chai'
 import hardhat, { ethers, deployments } from 'hardhat'
 import { AddAssetAction } from '../typechain-types'
-import { AddAssetHelper, STEP_TYPE_ID_ADD_ASSET } from '../tslib/helper'
+import { AddAssetHelper } from '../tslib/helper'
+import { STEP_TYPE_ID_ADD_ASSET } from '../../step-ids'
 import { EncodingContext } from '@freemarket/core'
 import { getTestFixture, MockWorkflowInstance, validateAction } from '@freemarket/step-sdk/tslib/testing'
 import { TestErc20__factory } from '@freemarket/step-sdk/typechain-types'
