+++ engine/steps/lido/deploy/01-deploy-deposit-eth.ts
@@ -1,6 +1,7 @@
 import { DeployFunction } from 'hardhat-deploy/types'
 import { deployStep } from '@freemarket/step-sdk/tslib/deploy-step'
-import { STEP_TYPE_ID_LIDO_ETH_TO_STETH, MAINNET_STETH_ADDRESS } from '../tslib'
+import { MAINNET_STETH_ADDRESS } from '../tslib'
+import { STEP_TYPE_ID_LIDO_ETH_TO_STETH } from '../../step-ids'
 
 const func: DeployFunction = async function (hardhatRuntimeEnv) {
   return deployStep('DepositEthForStEthAction', STEP_TYPE_ID_LIDO_ETH_TO_STETH, hardhatRuntimeEnv, [MAINNET_STETH_ADDRESS])
