+++ engine/steps/lido/test/eth-to-steth.ts
@@ -2,7 +2,8 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { expect } from 'chai'
 import hre, { ethers, deployments } from 'hardhat'
-import { MAINNET_STETH_ADDRESS, MAINNET_WSTETH_ADDRESS, STEP_TYPE_ID_LIDO_ETH_TO_STETH, STEP_TYPE_ID_LIDO_STETH_TO_WSTETH, STEP_TYPE_ID_LIDO_WSTETH_TO_STETH, encodeDepositEthForStEthParams, encodeWrapParams } from '../tslib'
+import { MAINNET_STETH_ADDRESS, MAINNET_WSTETH_ADDRESS, encodeDepositEthForStEthParams, encodeWrapParams } from '../tslib'
+import { STEP_TYPE_ID_LIDO_ETH_TO_STETH, STEP_TYPE_ID_LIDO_STETH_TO_WSTETH, STEP_TYPE_ID_LIDO_WSTETH_TO_STETH } from '../../step-ids'
 import { ADDRESS_ZERO, ASSET_TYPE_ERC20, ASSET_TYPE_NATIVE, AssetReference, createStandardProvider, EncodingContext, IERC20__factory, TEN_BIG } from '@freemarket/core'
 import { confirmTx, getTestFixture, getUsdt, MockWorkflowInstance, validateAction, WETH_ADDRESS } from '@freemarket/step-sdk/tslib/testing'
 import { WorkflowStruct } from '@freemarket/core/typechain-types/contracts/IWorkflowRunner'
