+++ engine/steps/lido/tslib/index.ts
@@ -1,6 +1,3 @@
 export * from './model'
 export const MAINNET_STETH_ADDRESS = '0xae7ab96520de3a18e5e111b5eaab095312d7fe84'
 export const MAINNET_WSTETH_ADDRESS = '0x7f39C581F595B53c5cb19bD0b3f8dA6c935E2Ca0'
-export const STEP_TYPE_ID_LIDO_ETH_TO_STETH = 205
-export const STEP_TYPE_ID_LIDO_STETH_TO_WSTETH = 206
-export const STEP_TYPE_ID_LIDO_WSTETH_TO_STETH = 207
