+++ engine/steps/pendle/tslib/index.ts
@@ -1,6 +1,5 @@
 export * from './model'
 
-export const STEP_TYPE_ID_PENDLE_TOKEN_TO_PT = 301
 
 export const MAINNET_PENDLE_ROUTER_STATIC = '0x263833d47eA3fA4a30f269323aba6a107f9eB14C'
 export const MAINNET_PENDLE_ROUTER = '0x888888888889758F76e7103c6CbF23ABbF58F946'
