+++ engine/steps/renzo/deploy/01-deploy-deposit-eth.ts
@@ -2,7 +2,8 @@ import { DeployFunction } from 'hardhat-deploy/types'
 import { getWrappedNativeAddress } from '@freemarket/step-sdk'
 import { deployStep } from '@freemarket/step-sdk/tslib/deploy-step'
 import { ADDRESS_ZERO } from '@freemarket/core'
-import { STEP_TYPE_ID_RENZO_ETH_TO_EZETH, MAINNET_EZETH_ADDRESS, MAINNET_RESTAKING_MANAGER_ADDRESS } from '../tslib'
+import { MAINNET_EZETH_ADDRESS, MAINNET_RESTAKING_MANAGER_ADDRESS } from '../tslib'
+import { STEP_TYPE_ID_RENZO_ETH_TO_EZETH } from '../../step-ids'
 
 const func: DeployFunction = async function (hardhatRuntimeEnv) {
   const chainId = await hardhatRuntimeEnv.getChainId()
