+++ engine/steps/renzo/test/eth-to-ezeth.ts
@@ -2,7 +2,8 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { expect } from 'chai'
 import hre, { ethers, deployments } from 'hardhat'
-import { MAINNET_EZETH_ADDRESS, STEP_TYPE_ID_RENZO_ETH_TO_EZETH } from '../tslib'
+import { MAINNET_EZETH_ADDRESS } from '../tslib'
+import { STEP_TYPE_ID_RENZO_ETH_TO_EZETH } from '../../step-ids'
 import { ADDRESS_ZERO, ASSET_TYPE_NATIVE, AssetReference, createStandardProvider, EncodingContext, IERC20__factory, TEN_BIG } from '@freemarket/core'
 import { confirmTx, getTestFixture, getUsdt, MockWorkflowInstance, validateAction, WETH_ADDRESS } from '@freemarket/step-sdk/tslib/testing'
 import { Weth__factory, formatNumber } from '@freemarket/step-sdk'
