+++ engine/steps/renzo/tslib/index.ts
@@ -1,4 +1,3 @@
 export * from './model'
 export const MAINNET_RESTAKING_MANAGER_ADDRESS = '0x74a09653A083691711cF8215a6ab074BB4e99ef5'
 export const MAINNET_EZETH_ADDRESS = '0xbf5495Efe5DB9ce00f80364C8B423567e58d2110'
-export const STEP_TYPE_ID_RENZO_ETH_TO_EZETH = 204
