+++ engine/steps/stargate-bridge/test/test.ts
@@ -3,7 +3,8 @@
 import { expect } from 'chai'
 import hardhat, { ethers, deployments } from 'hardhat'
 import { StargateBridgeAction } from '../typechain-types'
-import { StargateBridgeHelper, STEP_TYPE_ID_STARGATE_BRIDGE } from '../tslib/helper'
+import { StargateBridgeHelper } from '../tslib/helper'
+import { STEP_TYPE_ID_STARGATE_BRIDGE } from '../../step-ids'
 import { createStandardProvider, EncodingContext, TEN_BIG, WORKFLOW_END_STEP_ID, assert } from '@freemarket/core'
 import { confirmTx, getTestFixture, getUsdc, MockWorkflowInstance, validateAction } from '@freemarket/step-sdk/tslib/testing'
 import { getStargateComposerAddress } from '../tslib/stargateContractAddresses'
