+++ engine/steps/stargate-bridge/tslib/step-info.tsx
@@ -6,7 +6,7 @@ import AssetReferenceView from '@freemarket/step-sdk/build/tslib/helpers/AssetRe
 import { SummaryPropsBase } from '@freemarket/step-sdk'
 import { StargateBridge } from './model'
 import { capitalize } from '@freemarket/core'
-import { STEP_TYPE_ID_STARGATE_BRIDGE } from './helper'
+import { STEP_TYPE_ID_STARGATE_BRIDGE } from '../../step-ids'
 
 type Props = SummaryPropsBase<StargateBridge>
